frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
]
